snapshot of py3 dev

+++ bin/testEnvironment.py
@@ -8,6 +8,8 @@ print(tensorflow.__file__)
 print('importing DeepJetCore base')
 import DeepJetCore
 print(DeepJetCore.__file__)
+import os
+djc_base = os.environ.get('DEEPJETCORE')
 
 
 print('importing DJC masked tf.keras as keras...')
@@ -21,15 +23,19 @@ import numpy
 
 print('running random training in keras...')
 
-features = numpy.random.rand(1000, 10)
-truth = features
+def gen():
+    while(True):
+        features = numpy.random.rand(100, 10)
+        truth = features
+        yield (features,truth)
 
 
-a = Input(shape=(10,))
-b = keras.layers.Dense(10)(a)
+a = [Input(shape=(10,))]
+b = [keras.layers.Dense(10)(a[0])]
 model = keras.models.Model(inputs=a, outputs=b)
-model.compile(optimizer='adam',loss='mse')
-model.fit(x=features, y=truth, batch_size=100, epochs=3)
+model.compile(optimizer='adam',loss='mse',metrics=['accuracy'])
+model.fit(x=gen(), steps_per_epoch=100, batch_size=100, epochs=3)
+
 
 print('loading DeepJetCore compiled library...')
 
@@ -37,18 +43,18 @@ from DeepJetCore.compiled import c_arrayReads
 
 print('basic packages seem to be compiled... testing conversion')
 
-import os
-djc_base = os.environ.get('DEEPJETCORE')
-script='''
-#!/bin/bash
-cd {djc_base}/testing
-rm -rf batchDC 
-export PYTHONPATH=`pwd`:$PYTHONPATH
-convertFromSource.py -i files/filelist.txt -o batchDC -c TrainData_testBatch -n 1
-'''.format(djc_base=djc_base)
-os.system(script)
 
-print('testing batch explosion. Please check batch loss plot afterwards for smoothness. Warnings about the callback time can be ignored.')
+if True:
+    script='''
+    #!/bin/bash
+    cd {djc_base}/testing
+    rm -rf batchDC 
+    export PYTHONPATH=`pwd`:$PYTHONPATH
+    convertFromSource.py -i files/filelist.txt -o batchDC -c TrainData_testBatch -n 1
+    '''.format(djc_base=djc_base)
+    os.system(script)
+    
+    print('testing batch explosion. Please check batch loss plot afterwards for smoothness. Warnings about the callback time can be ignored.')
 
 script='''
 #!/bin/bash
